// apps/web/src/api/generated/models/identity/login400.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Talqo API
 * OpenAPI spec version: 0.0.1
 */
import * as zod from "zod"

export const Login400 = zod.union([
	zod.object({
		code: zod.literal("invalid-request"),
		type: zod.literal("https://docs.talqo.chat/problems#invalid-request"),
	}),
	zod.object({
		code: zod.literal("malformed-json"),
		type: zod.literal("https://docs.talqo.chat/problems#malformed-json"),
	}),
])

export type Login400 = zod.input<typeof Login400>
export type Login400Output = zod.output<typeof Login400>